services/config-service/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database ORM and driver
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Redis for caching
redis>=5.0.0

# Data validation and settings
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Kafka for event streaming
kafka-python>=2.0.2
aiokafka>=0.8.0

# Vault integration
hvac>=1.2.0

# YAML configuration parsing
pyyaml>=6.0

# ZooKeeper client
kazoo>=2.9.0

# HTTP client for health checks
aiohttp>=3.9.0

# Feature Flags - OpenFeature and Unleash
openfeature-sdk>=0.8.3
UnleashClient>=6.0.0

# Development dependencies (commented out for production)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
